+++ src/gamspy/_backend/backend.py
@@ -57,7 +57,7 @@ def backend_factory(
     if backend == "neos":
         from gamspy._backend.neos import NEOSServer
 
-        return NEOSServer(container, options, client, model)  # type: ignore
+        return NEOSServer(container, options, client, output, model)  # type: ignore
     elif backend == "engine":
         from gamspy._backend.engine import GAMSEngine
 
@@ -78,21 +78,19 @@ class Backend(ABC):
         self,
         container: Container,
         model: Model,
-        gdx_in: str,
-        gdx_out: str,
         options: Options,
+        output: io.TextIOWrapper | None,
     ):
         self.container = container
         self.model = model
-        self.gdx_in = gdx_in
-        self.gdx_out = gdx_out
         self.options = options
+        self.output = output
 
     @abstractmethod
     def is_async(self): ...
 
     @abstractmethod
-    def solve(self, keep_flags: bool = False): ...
+    def run(self, keep_flags: bool = False): ...
 
     def get_job_name(self):
         job_name = self.container._job
@@ -104,7 +102,7 @@ class Backend(ABC):
 
         return job_name
 
-    def preprocess(self, keep_flags: bool = False):
+    def preprocess(self, gdx_in: str, gdx_out: str, keep_flags: bool = False):
         (
             dirty_names,
             modified_names,
@@ -115,7 +113,7 @@ class Backend(ABC):
             self.container.write(self.container._gdx_in, modified_names)
 
         gams_string = self.container._generate_gams_string(
-            self.gdx_in, self.gdx_out, dirty_names, modified_names
+            gdx_in, gdx_out, dirty_names, modified_names
         )
 
         if not keep_flags:
@@ -123,7 +121,7 @@ class Backend(ABC):
 
         return gams_string, dirty_names
 
-    def prepare_summary(self, working_directory: str):
+    def prepare_summary(self, working_directory: str) -> pd.DataFrame:
         from gamspy._model import ModelStatus
 
         with open(