+++ tests/test_gamspy.py
@@ -45,7 +45,7 @@ class GamspySuite(unittest.TestCase):
     def test_version(self):
         import gamspy
 
-        self.assertEqual(gamspy.__version__, "0.13.0")
+        self.assertEqual(gamspy.__version__, "0.13.1")
 
 
 class DocsSuite(unittest.TestCase):