+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 GAMSPy CHANGELOG
 ================
 
+GAMSPy 0.13.1
+-------------
+- General
+  - Support output redirection for NEOS backend.
+  - Support GAMSPy to GAMS automatic conversion.
+  - Add support for old way of installing a license. 
+- Documentation
+  - Update model documentation to show how to redirect NEOS output to a file.
+- Testing
+  - Add a new model (knapsack) to the model library.
+
 GAMSPy 0.13.0
 -------------
 - General